delete question and answer use case

+++ src/domain/forum/application/usecases/delete-question.usecsae.ts
@@ -1,6 +1,7 @@
 import { QuestionRepository } from '../repositories/questions.repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -10,10 +11,13 @@ export class DeleteQuestionUseCase {
   constructor(private readonly questionRepository: QuestionRepository) {}
 
   async execute({
+    authorId,
     questionId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId)
     if (!question) throw new Error('Question not found')
+    if (authorId !== question.authorId.toString())
+      throw new Error('Not allowed')
     await this.questionRepository.delete(question)
     return {}
   }
